bootcamp/requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

asgiref==3.8.1
django==5.2.4
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
gunicorn==23.0.0
packaging==25.0
pyjwt==2.9.0
sqlparse==0.5.3
tzdata==2025.2; sys_platform == "win32"
